+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/engine/engine_router.test.tsx
@@ -17,6 +17,8 @@ jest.mock('../../../shared/flash_messages', () => ({
 }));
 import { setQueuedErrorMessage } from '../../../shared/flash_messages';
 
+import { Loading } from '../../../shared/loading';
+
 import { EngineRouter } from './';
 
 describe('EngineRouter', () => {
@@ -59,10 +61,10 @@ describe('EngineRouter', () => {
     );
   });
 
-  it('does not render if async data is still loading', () => {
+  it('renders a loading component if async data is still loading', () => {
     setMockValues({ ...values, dataLoading: true });
     const wrapper = shallow(<EngineRouter />);
-    expect(wrapper.isEmptyRender()).toBe(true);
+    expect(wrapper.find(Loading)).toHaveLength(1);
   });
 
   it('renders a default engine overview', () => {
