+++ x-pack/plugins/enterprise_search/public/applications/shared/loading/index.ts
