--- x-pack/plugins/enterprise_search/public/applications/workplace_search/components/shared/loading/loading.scss
@@ -1,14 +1,0 @@
-.loadingSpinnerWrapper {
-  width: 100%;
-  height: 90vh;
-  margin: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.loadingSpinner {
-  width: $euiSizeXXL * 1.25;
-  height: $euiSizeXXL * 1.25;
-}
